refactor(workspace): extract useWorkspaceConversations hook (1956→1774 lines)
Moves conversation management, intake triage lookup, accepted-intake fetching,
and filteredConversations derivation into a dedicated hook.
